+++ src/Libraries/Microsoft.Extensions.Diagnostics.Extra/Enrichment/ProcessEnricherServiceCollectionExtensions.cs
@@ -12,7 +12,7 @@ namespace Microsoft.Extensions.DependencyInjection;
 /// <summary>
 /// Provides extension methods for setting up Process enrichers in an <see cref="IServiceCollection" />.
 /// </summary>
-public static class ProcessEnricherExtensions
+public static class ProcessEnricherServiceCollectionExtensions
 {
     /// <summary>
     /// Adds an instance of the process enricher to the <see cref="IServiceCollection"/>.
